moved missing hdrp code to hdrp
